requirements.txt
asgiref==3.6.0
black==22.12.0
click==8.1.3
Django==4.1.5
django-filter==22.1
django-phonenumber-field==7.0.2
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mypy-extensions==0.4.3
pathspec==0.10.3
phonenumberslite==8.13.4
Pillow==9.4.0
platformdirs==2.6.2
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
